backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
minio
Pillow
python-dotenv
fastapi-cors
passlib[bcrypt]
python-jose